Added to API, core, Infrastructure, persistence
ValidationFilter template added, removed kube configuration seeing as it was obsolotte. Added mongoDB configuration instead. Made kubectl deployment and service in one file for ease of use.  Added eventPublisher to mediator. Created base mongoDB repository class and interface. As well as a few methods in the repository. Added the first repository which is ProcessesRepository